Remove warning in elixir 1.8: due to for :into with non empty list
It fixes the following warning
warning: the Collectable protocol is deprecated for non-empty lists. The
behaviour of things like Enum.into/2 or "for" comprehensions with an
:into option is incorrect when collecting into non-empty lists. If
you're collecting into a non-empty keyword list, consider using
Keyword.merge/2 instead. If you're collecting into a non-empty list,
consider concatenating the two lists with the ++ operator.
  (elixir) lib/collectable.ex:83: Collectable.List.into/1
    (stdlib) erl_eval.erl:680: :erl_eval.do_apply/6
      (stdlib) erl_eval.erl:449: :erl_eval.expr/5
        (stdlib) erl_eval.erl:126: :erl_eval.exprs/5
          (elixir) src/elixir.erl:258: :elixir.eval_forms/4